+++ Changelog.md
@@ -2,11 +2,25 @@
 
 ### Ignition Gazebo 2.XX.X (201X-XX-XX)
 
+### Ignition Gazebo 2.13.0 (2019-12-17)
+
+1. Add Multicopter velocity controller
+    * [Pull Request 487](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/487)
+
 1. Fix crash when removing an entity being followed
     * [Pull Request 465](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/465)
 
 1. Add option to right click and remove nodes
     * [Pull Request 458](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/458)
+
+1. Fix jumpy log playback
+    * [Pull Request 488](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/488)
+
+1. Remove Scene3d Text anchors
+    * [Pull Request 467](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/467)
+
+1. Show grid using SDF file
+    * [Pull Request 461](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/461)
 
 ### Ignition Gazebo 2.12.0 (2019-11-25)